# front/src/api/models/abstract-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AbstractResponse
 */
export interface AbstractResponse {
    /**
     * 
     * @type {string}
     * @memberof AbstractResponse
     */
    'title'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AbstractResponse
     */
    'message'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AbstractResponse
     */
    'status'?: string | null;
    /**
     * Произвольные данные ответа
     * @type {any}
     * @memberof AbstractResponse
     */
    'data'?: any | null;
}