# front/src/api/models/sell-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface SellRequestDTO
 */
export interface SellRequestDTO {
    /**
     * 
     * @type {number}
     * @memberof SellRequestDTO
     */
    'ticketId': number;
    /**
     * 
     * @type {number}
     * @memberof SellRequestDTO
     */
    'personId': number;
    /**
     * 
     * @type {number}
     * @memberof SellRequestDTO
     */
    'amount': number;
}