# front/src/api/models/venue.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tickets API
 * REST API для управления билетами, людьми, событиями и площадками.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Venue
 */
export interface Venue {
    /**
     * 
     * @type {number}
     * @memberof Venue
     */
    'id'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Venue
     */
    'name'?: string;
    /**
     * 
     * @type {number}
     * @memberof Venue
     */
    'capacity'?: number | null;
    /**
     * 
     * @type {string}
     * @memberof Venue
     */
    'type'?: VenueTypeEnum | null;
}

export const VenueTypeEnum = {
    LOFT: 'LOFT',
    OPEN_AREA: 'OPEN_AREA',
    STADIUM: 'STADIUM'
} as const;

export type VenueTypeEnum = typeof VenueTypeEnum[keyof typeof VenueTypeEnum];